+++ components/home/Features.tsx
@@ -1,0 +1,5 @@
+export default function Features() {
+  return (
+    <>Features</>
+  )
+}
